refactor: 전체 코드 정리
