+++ src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,0 +1,15 @@
+package roomescape.dto;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import roomescape.domain.Reservation;
+
+public record ReservationResponse(Long id, String name, LocalDate date, LocalTime time) {
+    public static ReservationResponse from(Reservation reservation) {
+        return new ReservationResponse(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                reservation.getTime());
+    }
+}
